Beginning testing implementations

+++ kollekt/__init__.py
@@ -3,6 +3,8 @@ from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 from flask_login import LoginManager
 
+import config
+
 db = SQLAlchemy()
 login_manager = LoginManager()
 
@@ -17,8 +19,9 @@ def create_app():
 
     with app.app_context():
         from . import routes
-        db.drop_all()
-        print("Dropped All Tables")
+        if CONFIG_TYPE == config.DevelopmentConfig:
+            db.drop_all()
+            print("Dropped All Tables")
         db.create_all()
 
         print("Created All Tables From Models")
